Move to BSC chain deployment
