+++ .env.example
@@ -1,10 +1,10 @@
 PRIVATE_KEY_DEPLOY=0x...
 
-CONSTANTS_FILENAME="Blast.json"
-OUTPUT_FILENAME="Blast-Main.json"
+CONSTANTS_FILENAME="Bsc.json"
+OUTPUT_FILENAME="Bsc-Main.json"
 
-BLAST_RPC_URL="https://blast-mainnet.g.alchemy.com/v2/zBnxi7r46b-Kp-L3OqbKeC_loHcfj4vc"
-BLAST_SCAN_API_KEY=https://api.blastscan.io/api
-BLAST_ETHERSCAN_VERIFIER_URL=https://api.blastscan.io/api
+BSC_RPC_URL="https://bsc-mainnet.g.alchemy.com/v2/zBnxi7r46b-Kp-L3OqbKeC_loHcfj4vc"
+BSC_SCAN_API_KEY=https://api.bscscan.com/api
+BSC_ETHERSCAN_VERIFIER_URL=https://api.bscscan.com/api
 
 FORK_BLOCK_NUMBER=0
