+++ contracts/VeArtProxy.sol
@@ -81,7 +81,7 @@ contract VeArtProxy is IVeArtProxy {
                     abi.encodePacked(
 			            '{"name": "lock #',
                         toString(cfg._tokenId),
-                        '", "background_color": "121a26", "description": "Tenex Finance is a next-generation AMM that combines the best of Curve, Solidly and Uniswap, designed to serve as the Blast central liquidity hub. Tenex NFTs vote on token emissions and receive bribes and fees generated by the protocol.", "image": "data:image/svg+xml;base64,',
+                        '", "background_color": "121a26", "description": "Tenex Finance is a next-generation AMM that combines the best of Curve, Solidly and Uniswap, designed to serve as the Bsc central liquidity hub. Tenex NFTs vote on token emissions and receive bribes and fees generated by the protocol.", "image": "data:image/svg+xml;base64,',
                         Base64.encode(bytes(output)), '", ',
                         '"attributes": [', attributes, ']',
 			            '}'
