+++ foundry.toml
@@ -6,12 +6,12 @@ fs_permissions = [{ access = "read-write", path = "./"}]
 
 [rpc_endpoints]
 goerli = "${GOERLI_RPC_URL}"
-blast = "${BLAST_RPC_URL}"
+bsc = "${BSC_RPC_URL}"
 
 
 [etherscan]
 goerli = { key = "${GOERLI_SCAN_API_KEY}", url = "${GOERLI_ETHERSCAN_VERIFIER_URL}" }
-blast = { key = "${BLAST_SCAN_API_KEY}", url = "${BLAST_ETHERSCAN_VERIFIER_URL}" }
+bsc = { key = "${BSC_SCAN_API_KEY}", url = "${BSC_ETHERSCAN_VERIFIER_URL}" }
 
 [invariant]
 runs = 2000
