+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "tenex",
   "version": "2.0.0",
-  "description": "Base layer AMM on Blast",
+  "description": "Base layer AMM on Bsc",
   "directories": {
     "lib": "lib"
   },
@@ -10,12 +10,12 @@
     "@nomicfoundation/hardhat-foundry": "^1.0.1",
     "@nomicfoundation/hardhat-network-helpers": "^1.0.8",
     "@nomicfoundation/hardhat-toolbox": "^2.0.2",
-    "@nomiclabs/hardhat-etherscan": "^3.1.7",
     "@openzeppelin/contracts": "4.8.0",
     "@tenderly/hardhat-tenderly": "1.7.5",
     "@typechain/ethers-v5": "^10.2.1",
     "@typechain/hardhat": "^6.1.6",
     "@types/chai": "^4.3.5",
+    "@types/minimatch": "^5.1.2",
     "@types/mocha": "^10.0.1",
     "chai": "^4.3.7",
     "dotenv": "^16.0.3",
@@ -28,7 +28,7 @@
     "solhint-plugin-prettier": "^0.0.5",
     "solidity-coverage": "^0.8.2",
     "ts-node": "^10.9.1",
-    "typechain": "^8.2.0",
+    "typechain": "^8.3.2",
     "typescript": "^5.1.3"
   },
   "scripts": {
@@ -38,6 +38,7 @@
     "lint:check": "solhint 'contracts/**/*.sol' 'test/**/*.sol' 'script/**/*.sol'"
   },
   "dependencies": {
-    "forge": "^2.3.0"
+    "@nomicfoundation/hardhat-verify": "^2.1.1",
+    "nvm": "^0.0.4"
   }
 }
