--- script/constants/BlastSepolia.json
@@ -1,34 +1,0 @@
-{
-    "feeManager": "0x3c5a809e712D30D932b71EdB066FA2EEDEE6Ad58",
-    "allowedManager": "0x3c5a809e712D30D932b71EdB066FA2EEDEE6Ad58",
-    "emergencyCouncil": "0x3c5a809e712D30D932b71EdB066FA2EEDEE6Ad58",
-    "team": "0x3c5a809e712D30D932b71EdB066FA2EEDEE6Ad58",
-    "vetoer": "0x3c5a809e712D30D932b71EdB066FA2EEDEE6Ad58",
-    "WETH": "0x4200000000000000000000000000000000000023",
-    "TENEX": "0x8Fe5C378B39DBdC984F2FfB8631227758340631A",
-    "merkleRoot": "0xbd6c68af958c9ab9df815558f13d83a60119d22cd81df06da12f49da30481ec8",
-    "whitelistTokens": [
-        "0x4200000000000000000000000000000000000022",
-        "0xD88957c98D65E9bee30304290f734847De09B990"
-    ],
-    "poolsV2": [
-        {
-            "stable": false,
-            "tokenA": "0x4200000000000000000000000000000000000022",
-            "tokenB": "0xD88957c98D65E9bee30304290f734847De09B990"
-        }
-    ],
-    "poolsTenex": [
-        {
-            "stable": false,
-            "token": "0x4200000000000000000000000000000000000022"
-        }
-    ],
-    
-    "current": {
-        "Forwarder": "0x84A4A7254bB4aee5F1404BB7d793d285EbF4E580",
-        "Minter": "0x7580e59EdDC15d5AB9Ccd880Aa8bd1F50d399Ab8",
-        "VotingEscrow": "0xF1da90aE93B0f8749f4e7CCAE18B8Cb347fD8Ac8",
-        "Voter": "0xcC5821395D8Aba8DFb89D55a99ce724cBeab1bee"
-    }
-}
