--- script/constants/Optimism.json
@@ -1,42 +1,0 @@
-{
-    "allowedManager": "",
-    "feeManager": "",
-    "emergencyCouncil": "",
-    "team": "",
-    "vetoer": "",
-    "WETH": "",
-    "whitelistTokens": [
-        
-    ],
-    "poolsV2": [
-        {
-            "stable": true,
-            "tokenA": "",
-            "tokenB": ""
-        }
-       
-    ],
-    "poolsTenexV2": [
-        {
-            "stable": true,
-            "tokenB": ""
-        }
-       
-    ],
-    "v1": {
-        "TENEX": "",
-        "VotingEscrow": "",
-        "Voter": "",
-        "Factory": "",
-        "Router": "",
-        "Gov": "",
-        "Distributor": "",
-        "Minter": ""
-    },
-    "current": {
-        "Forwarder": "",
-        "Minter": "",
-        "VotingEscrow": "",
-        "Voter": ""
-    }
-}
