--- script/constants/contract-addr/DeployGaugesAndPools-Blast-Test.json
@@ -1,10 +1,0 @@
-{
-  "gaugesPoolsV2": [
-    "0x1f6f5f0F23EaB7dA113418aaa8918eCb3b7f9d13",
-    "0xE47207888eF6466Bc0322204199498E8470e5210"
-  ],
-  "poolsV2": [
-    "0xB1B390Ff7ec46Eae81A8E3463E2EdFBEC88D7C43",
-    "0x7629Da077C82F2c75d9d33ec10E527450c3C7137"
-  ]
-}
