--- script/constants/contract-addr/DeployGovernors-Blast-Test.json
@@ -1,4 +1,0 @@
-{
-  "EpochGovernor": "0xbB5F87eA61F6c155c74608CD8Dd19d4Ab23bFeA1",
-  "Governor": "0xa534546309062D37c9a32A67a466179b0e2975F3"
-}
