--- script/constants/contract-addr/TenexOutput.json
@@ -1,15 +1,0 @@
-{
-  "artProxy": "0x8d6C181b8001cE35f7E7020e0390B45CC0800F6f",
-  "distributor": "0x6E82043015165cacf2e725E00B31583Fb14b7c4d",
-  "factoryRegistry": "0x884285225B0C6b83a205046BD5e016e472d63450",
-  "forwarder": "0x259853810e540078De1766B19B90c4E0143C6351",
-  "gaugeFactory": "0x89EDeECf49c54F2d91521f10061ef257928Cf1f5",
-  "managedRewardsFactory": "0xB8e20A9048A9B0743E769340D0e95F2DC05C461e",
-  "minter": "0x43976AEF56B6f0Ef083D41134640403FCF770405",
-  "poolFactory": "0xF7Fa6B336c45d354c7891123b3FffBf0eC864614",
-  "router": "0x05fF21a1a50515165dA4F88fE96e573b7D08D6dB",
-  "TENEX": "0x50BA9B7E05cE03f1a9a5D154AA4fb1523d5D70Ed",
-  "voter": "0x6Ea547BC0f97A28b8eF6F3dB99Cf85E27F6b66fb",
-  "votingEscrow": "0x32046405e9fc0eE94bDfCED242fa99Ee7A589ea0",
-  "votingRewardsFactory": "0xAc77aC4f5C9A51d50928d9BdA979a9F1e9B82d24"
-}
