--- script/constants/contract-addr/TenexOutput2.json
@@ -1,15 +1,0 @@
-{
-  "artProxy": "0xc35d7cda09a239d38b4587F999b57B024E8cdF2C",
-  "distributor": "0x01BbAE08E0B213ba382d7b36433ef13C8F58Fda6",
-  "factoryRegistry": "0xf4b1e32B897381eB3C9bF0457B77c58A51B6D5b1",
-  "forwarder": "0x13E6D2f8bE44b02Dd0D5924d4A63a46B0efFCB41",
-  "gaugeFactory": "0x9eDaed16E47B5d7706Eda74Eb2EbE71Ff05693DB",
-  "managedRewardsFactory": "0x243b986437b72D63DFed01204BB9E2D24c657b5b",
-  "minter": "0xD0754fbA9a35723acBdE561443ae9EBbDbd6fB56",
-  "poolFactory": "0x8aa67e3468862a1b46c24A2B2D68F69fb187bafB",
-  "router": "0x56c3EcAb9A6bF1e65fF67F66C69DE4D5FBdB9f1A",
-  "TENEX": "0x0D9915280057BA01a2EaAe6a5B14A10D6860e377",
-  "voter": "0xC64be50B7fb03c00d3332e4E2392fFe00CD2F637",
-  "votingEscrow": "0x412064ec3EC4F868A89C2cfDE3Aed58531b7C34B",
-  "votingRewardsFactory": "0x6703bB0f2C9816fff9678445c3309F98844D3BDe"
-}
