--- script/constants/contract-addr/TenexOutputPools2.json
@@ -1,14 +1,0 @@
-{
-  "poolsV2": [
-    "0x1596b13A748d339c02d58c60cA25710E139aC96c",
-    "0x63Bc75c581069740095B96A5559eaD3cfcf45cda"
-  ],
-  "gaugesPoolsV2": [
-    [
-      "0x4aDB3aB229896ACECb52dC31aed6AC8868D85aB3"
-    ],
-    [
-      "0xc93e188727213E5F432d77FEE0e3441D9b53d976"
-    ]
-  ]
-}
