--- script/constants/output/DeployGaugesAndPools-Blast-Main.json
@@ -1,10 +1,0 @@
-{
-  "gaugesPoolsV2": [
-    "0x35CD092170CABAd26F9B65b33757B07670CdCDC7",
-    "0xC924Fdb64f9d82ffE5f8a76b310AC5565fB2cf5d"
-  ],
-  "poolsV2": [
-    "0x794af1d173807dAb1aC6e1c46EA6927A361209Bf",
-    "0x1D95F3f2Efb0a56Cd35e315bb74BD4AeEf2198A1"
-  ]
-}
