--- script/constants/output/DeployGovernors-Blast-Main.json
@@ -1,4 +1,0 @@
-{
-  "EpochGovernor": "0x7B650f81252e4d805a37f02efc92a4A1fCc8bcC8",
-  "Governor": "0xe525E6E0c6132D12ADD108dd0BB594F515dd6498"
-}
