--- script/constants/output/DeployTenex-Blast-Main.json
@@ -1,15 +1,0 @@
-{
-  "ArtProxy": "0x766B1c39428785fE830fcb0cf81dbFdEDcd866d5",
-  "Distributor": "0x866439457B98dEB907fA038728E36dB7aefEE794",
-  "FactoryRegistry": "0x09B40a2d9091B985849AAf86EebFDcad3F106268",
-  "Forwarder": "0x414Cf302F001a3Cff313EbfA07C317Fe548D03ad",
-  "GaugeFactory": "0xa2bac2F3204C5E3734435E02208443Eb68f62f6e",
-  "ManagedRewardsFactory": "0x84c4D31A76acbf83CBAD7A03e019625432dcfF83",
-  "Minter": "0xe20086E01FF1382332cf72c53D2bF5014bA1fBbe",
-  "PoolFactory": "0x9FC1091Bf704F2a74399E2ae98517202cab821b6",
-  "Router": "0x6e0EaF9D4007410F9053cCF3494847e2de64626a",
-  "TENEX": "0xC4B7846887648B4bef80db245411B0323Edd2a66",
-  "Voter": "0xDaE876Af086d7E2ee2CF5434faAB6DD54732BcA5",
-  "VotingEscrow": "0xfec8C7a0129cFCad689682C972b817C549f593b6",
-  "VotingRewardsFactory": "0x83CCe00595b74D47Eb4872da4Eae36AaD1E557C2"
-}
